requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
Flask==2.2.3
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openai==1.1.1
psycopg2-binary==2.9.9
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.0.0
regex==2024.11.6
requests==2.28.2
slack-sdk==3.20.2
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typesense==1.0.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
Werkzeug==2.2.3
zipp==3.21.0
